improved update script

+++ _update_locally.sh
@@ -1,7 +1,11 @@
 # Only use this script while developing!
 # Exit Kodi, run this script and start Kodi again.
-rm -rf ~/.kodi/addons/plugin.video.intergalacticfm/*
-cp -a ../plugin.video.intergalacticfm/* ~/.kodi/addons/plugin.video.intergalacticfm/
+if [ -e ~/.kodi/addons/plugin.video.intergalacticfm ]; then
+	rm -rf ~/.kodi/addons/plugin.video.intergalacticfm/*
+	cp -a ../plugin.video.intergalacticfm/* ~/.kodi/addons/plugin.video.intergalacticfm/
+fi
 #
-rm -rf ~/.kodi/addons/plugin.audio.intergalacticfm/*
-cp -a ../plugin.audio.intergalacticfm/* ~/.kodi/addons/plugin.audio.intergalacticfm/
+if [ -e ~/.kodi/addons/plugin.audio.intergalacticfm ]; then
+	rm -rf ~/.kodi/addons/plugin.audio.intergalacticfm/*
+	cp -a ../plugin.audio.intergalacticfm/* ~/.kodi/addons/plugin.audio.intergalacticfm/
+fi
